Split automation retry orchestration
